Partners Tab

+++ Core/src/com/mfino/dao/PartnerDAO.java
@@ -95,7 +95,7 @@ public class PartnerDAO extends BaseDAO<Partner> {
 								CmFinoFIX.TagID_BusinessPartnerTypePartner)) {
 					DetachedCriteria partner = DetachedCriteria.forClass(EnumText.class).
 												setProjection(Property.forName(EnumText.FieldName_EnumCode))
-												.add(Restrictions.eq(EnumText.FieldName_TagID, query.getPartnerType()));
+												.add(Restrictions.eq(EnumText.FieldName_TagID, query.getPartnerType().longValue()));
 //					HashMap<String, String> results = enumTextService
 //							.getEnumTextSet(query.getPartnerType(), null);
 //					if (!results.isEmpty()) {
